requirements.txt
flask
sentence-transformers
pinecone-client
google-generativeai
